Clean up data_transfer_objects
