--- nucleus/modelci/data_transfer_objects/get_eval_functions.py
@@ -1,9 +1,0 @@
-from typing import List
-
-from pydantic import BaseModel
-
-from ..eval_function import EvalFunctionEntry
-
-
-class GetEvalFunctions(BaseModel):
-    eval_functions: List[EvalFunctionEntry]
