+++ nucleus/modelci/unit_test.py
@@ -7,7 +7,7 @@ and have confidence that they’re always shipping the best model.
 from dataclasses import dataclass
 from typing import List, TYPE_CHECKING
 
-from .data_transfer_objects.add_unit_test_metric import AddUnitTestMetric
+from .data_transfer_objects.unit_test_metric import AddUnitTestMetric
 from .data_transfer_objects.unit_test_evaluations import GetEvalHistory
 from .unit_test_evaluation import UnitTestEvaluation
 from .unit_test_metric import UnitTestMetric
